+++ source/commands/init/default-dangerfile.ts
@@ -1,0 +1,5 @@
+import { InitState } from "../danger-init"
+
+export const generateDefaultDangerfile = (_state: InitState) => {
+  return "import {danger, etc}"
+}